use local archives if available instead of downloading them

+++ run.sh
@@ -1,22 +1,37 @@
 #!/bin/sh
 
-die() { echo; echo "# [ERROR] see folder '$FOLDER' and '$( pwd )' - aborting"; exit 1; }
+die()
+{
+	echo
+	echo "# [ERROR] see folder '$FOLDER' and '$( pwd )' - aborting"
+	exit 1
+}
 
+download()
+{
+	BASE="$( basename "$URL" )"
+
+	if [ -f "$GITBASE/$BASE" ]; then
+		cp -v "$GITBASE/$BASE" "$BASE" || die
+	else
+		wget --no-check-certificate -O "$BASE" "$URL" || die
+	fi
+}
+
+GITBASE="$( pwd )"
 FOLDER="$( mktemp -d )" || die
 cd "$FOLDER" || die
 echo "# [OK] working in folder: '$FOLDER'"
 
 # URL='http://deb.debian.org/debian/pool/main/s/slirp/slirp_1.0.17.orig.tar.gz'
 URL='https://github.com/bittorf/slirp-uml-and-compiler-friendly/raw/main/slirp_1.0.17.orig.tar.gz'
-BASE="$( basename "$URL" )"
-wget --no-check-certificate -O "$BASE" "$URL" || die
+download
 tar xzf "$BASE"			# folder 'slirp_1.0.17'
 
 # URL='http://deb.debian.org/debian/pool/main/s/slirp/slirp_1.0.17-11.debian.tar.xz'
 URL='https://github.com/bittorf/slirp-uml-and-compiler-friendly/raw/main/slirp_1.0.17-11.debian.tar.gz'
-BASE="$( basename "$URL" )"
-wget --no-check-certificate -O "$BASE" "$URL" || die
-tar xJf "$BASE"			# folder 'debian'
+download
+tar xzf "$BASE"			# folder 'debian'
 
 cd ./*.17 || die
 RC=0
